:recycle: Separate file for SectionHeaderFooter
