+++ Conv/SectionHeaderFooter/SectionHeaderFooter+Delegate.swift
@@ -1,0 +1,37 @@
+//
+//  SectionHeaderFooter+Delegate.swift
+//  Conv
+//
+//  Created by Yudai.Hirose on 2018/07/25.
+//  Copyright © 2018年 廣瀬雄大. All rights reserved.
+//
+
+import UIKit
+
+extension SectionHeaderFooter: SectionHeaderFooterDelegatable {
+    public func configureView(_ collectionView: UICollectionView, view: UICollectionReusableView, section: Int) {
+        guard let view = view as? View else {
+            fatalError()
+        }
+        configureView?(view, (self, collectionView, section))
+    }
+    
+    public func sizeFor(_ collectionView: UICollectionView, section: Int) -> CGSize? {
+        return sizeFor?((self, collectionView, section)) ?? size
+    }
+    
+    public func willDisplay(_ collectionView: UICollectionView, view: UICollectionReusableView, indexPath: IndexPath) {
+        guard let view = view as? View else {
+            fatalError()
+        }
+        willDisplay?(view, (self, collectionView, indexPath))
+    }
+    
+    public func didEndDisplay(_ collectionView: UICollectionView, view: UICollectionReusableView, indexPath: IndexPath) {
+        guard let view = view as? View else {
+            fatalError()
+        }
+        didEndDisplay?(view, (self, collectionView, indexPath))
+    }
+}
+
